Cleanup. Added draw.c.
